requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
alembic==1.12.0
Flask-JWT-Extended==4.5.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0